requirements.txt
requests>=2.28
psycopg2-binary>=2.9
python-dotenv>=1.0
